requirements.txt
confluent-kafka==2.5.3
yfinance==0.2.54
psycopg[binary]==3.2.1
great-expectations==1.2.4
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
